# clients/typescript/openapi.json
{
  "components": {
    "schemas": {
      "CompletionRequest": {
        "properties": {
          "engine": {
            "description": "Completions engine token, e.g. 'openai', 'claude', 'google-gemini'.",
            "title": "Engine",
            "type": "string"
          },
          "max_response_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Response Tokens"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model name; omit to use the engine's default model.",
            "title": "Model"
          },
          "prompt": {
            "maxLength": 200000,
            "title": "Prompt",
            "type": "string"
          },
          "request_timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Timeout Seconds"
          },
          "stream": {
            "default": false,
            "description": "When true the response is text/event-stream (SSE). Streaming responses bypass PII redaction middleware by library design.",
            "title": "Stream",
            "type": "boolean"
          },
          "system_prompt": {
            "anyOf": [
              {
                "maxLength": 50000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "required": [
          "engine",
          "prompt"
        ],
        "title": "CompletionRequest",
        "type": "object"
      },
      "CompletionResponse": {
        "description": "Buffered completion result.\n\nText only, because `asend_prompt` returns a bare `str`. Token usage and a\nfinish reason would require routing this through the conversation call\ninstead, which is a different provider path than the endpoint documents.\n`/v1/conversations/turn` and `/v1/structured` return both natively.",
        "properties": {
          "engine": {
            "description": "Engine that served the request.",
            "title": "Engine",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model requested, or null for the engine default.",
            "title": "Model"
          },
          "text": {
            "description": "Generated completion text.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "engine"
        ],
        "title": "CompletionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP interface to the ai-api-unified Python library. Breaking API changes bump the URI version (/v1 -> /v2); the service itself follows semantic versioning independently.",
    "title": "ai-api-unified-http",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/completions": {
      "post": {
        "description": "Generate a completion, buffered by default or streamed over SSE.\n\nArgs:\n    request: Engine selection, prompt, and generation options.\n\nReturns:\n    CompletionResponse | StreamingResponse: The completion text, or an SSE\n        stream when `stream` is true.",
        "operationId": "completions_v1_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            },
            "description": "Buffered completion. With stream=true the response is instead text/event-stream carrying chunk events, then one terminal done or error event."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Text completion (buffered or SSE stream)"
      }
    },
    "/v1/conversations/turn": {
      "post": {
        "description": "Run one assistant turn; the caller owns history and tool execution.\n\nArgs:\n    request: System prompt, full history, optional tools, and the previous\n        turn's opaque token.\n\nReturns:\n    ConversationTurnResponse: The assistant turn, any tool calls the caller\n        must execute, and the token to echo back next turn.",
        "operationId": "conversation_turn_v1_conversations_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "One stateless conversation turn, with optional tool schemas"
      }
    },
    "/v1/embeddings": {
      "post": {
        "description": "Embed one or more inputs.\n\nArgs:\n    request: Engine selection and the texts to embed.\n\nReturns:\n    EmbeddingsResponse: One vector per input, index-paired to the inputs.",
        "operationId": "embeddings_v1_embeddings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Embedding vectors for one or more inputs"
      }
    },
    "/v1/models": {
      "get": {
        "description": "List an engine's models with their lifecycle and pricing.\n\n`engine` is required rather than optional. Listing every engine would mean\nconstructing a client for each one on a single request, and client\nconstruction re-reads configuration and makes a network round trip on\nGemini. One engine per call keeps the cost of this endpoint predictable.\n\nArgs:\n    engine: Engine token to list.\n\nReturns:\n    ModelsResponse: Provider-reported models, plus registry entries for\n        lifecycle and pricing.",
        "operationId": "models_v1_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "engine",
            "required": true,
            "schema": {
              "title": "Engine",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Catalogued models with capabilities, pricing, and lifecycle"
      }
    },
    "/v1/structured": {
      "post": {
        "description": "Generate output validated against a caller-supplied JSON schema.\n\nArgs:\n    request: Engine selection, prompt or messages, and the response schema.\n\nReturns:\n    StructuredResponse: Parsed data with the finish reason and usage. Data\n        is null on any finish reason other than `complete`.",
        "operationId": "structured_v1_structured_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructuredResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Schema-validated structured output"
      }
    },
...
  }
}